Update ResponseAggregationsConverter name to AggregationsConverter
